+++ src/frontend/apps/bbb/DashboardMeetingJoin/index.tsx
@@ -1,13 +1,8 @@
-import { Anchor, Box, Paragraph } from 'grommet';
+import { Anchor, Box } from 'grommet';
 import React from 'react';
-import { defineMessages, FormattedMessage, useIntl } from 'react-intl';
+import { defineMessages, useIntl } from 'react-intl';
 
 const messages = defineMessages({
-  blockedPopups: {
-    defaultMessage: 'It seems like your browser is blocking popups.',
-    description: 'Message when browser is blocking popups.',
-    id: 'component.DashboardMeetingJoin.blockedPopups',
-  },
   joinLinkLabel: {
     defaultMessage: 'Please click here to access meeting.',
     description: 'Label for link used to join a meeting.',
@@ -17,7 +12,7 @@ const messages = defineMessages({
 
 interface DashboardMeetingJoinProps {
   href: string;
-  onClick: () => void;
+  onClick?: () => void;
 }
 const DashboardMeetingJoin = ({ href, onClick }: DashboardMeetingJoinProps) => {
   const intl = useIntl();
@@ -31,9 +26,6 @@ const DashboardMeetingJoin = ({ href, onClick }: DashboardMeetingJoinProps) => {
       border={{ color: 'accent-1', size: 'small' }}
       background={{ color: 'accent-1', opacity: 'weak' }}
     >
-      <Paragraph margin="none">
-        <FormattedMessage {...messages.blockedPopups} />
-      </Paragraph>
       <Anchor
         href={href}
         target="_blank"